separate utils functions to handle all request types
